requirements.txt
flask
line-bot-sdk<3.0
google-generativeai
gunicorn
python-dotenv